requirements.txt
flask
Pillow
vtracer
